Refactor the staking calculator mechanism
The previous implementation for the staking rewards calculation was flawed, this refactoring is an improvement but still doesn't take sell transactions into account since they aren't supported yet by the portfolio.

The current implementation will loop the LSD coins and loop their transactions. For each transaction will the difference in conversion rate be calculated to know what the gain (or loss) from staking is. The staking rewards will be summed up in both ETH and USD.